Extract CommandApplicationProcess from CoreState.
CoreState now takes over storeId checking from BatchingMessageHandler.
