+++ enterprise/core-edge/src/main/java/org/neo4j/coreedge/raft/RaftStateMachine.java
@@ -41,4 +41,6 @@ public interface RaftStateMachine
     void notifyNeedFreshSnapshot();
 
     void downloadSnapshot( MemberId from );
+
+    void innerHandle( RaftMessages.StoreIdAwareMessage raftMessage );
 }
